# Use builtin :crypto functions (when available) for PBKDF2 and secure constant-time comparison.